refactor: change icon set
