+++ packages/client/src/App.tsx
@@ -8,30 +8,35 @@ import { useGetMe } from './hooks/queries/user';
 import Home from './pages/Home/Home';
 import Search from './pages/Search/Search';
 import Write from './pages/Write/Write';
-import Bookmarks from './pages/Bookmarks/Bookmarks';
+import Room from './pages/Room/Room';
 import Settings from './pages/Settings/Settings';
-import Chat from './pages/Room/Room';
+import Chat from './pages/Room/Chat';
 import NotFound from './pages/NotFound/NotFound';
+import ErrorBoundary from './components/base/ErrorBoundary';
+import ErrorFallback from './components/base/ErrorFallback';
+import { MESSAGE } from './constants';
 
 const App = () => {
   useGetMe();
 
   return (
-    <Router>
-      <Routes>
-        {/* public routes */}
-        <Route path="/" element={<Home />} />
-        <Route path="/search" element={<Search />} />
-        <Route path="/write" element={<Write />} />
-        <Route path="/bookmarks" element={<Bookmarks />} />
-        <Route path="/settings" element={<Settings />} />
+    <ErrorBoundary fallback={<ErrorFallback message={MESSAGE.ERROR.UNKNOWN} />}>
+      <Router>
+        <Routes>
+          {/* public routes */}
+          <Route path="/" element={<Home />} />
+          <Route path="/search" element={<Search />} />
+          <Route path="/write" element={<Write />} />
+          <Route path="/room" element={<Room />} />
+          <Route path="/setting" element={<Settings />} />
 
-        <Route path="/room/:roomId" element={<Chat />} />
+          <Route path="/room/:roomId" element={<Chat />} />
 
-        {/* catch all */}
-        <Route path="*" element={<NotFound />} />
-      </Routes>
-    </Router>
+          {/* catch all */}
+          <Route path="*" element={<NotFound />} />
+        </Routes>
+      </Router>
+    </ErrorBoundary>
   );
 };
