+++ packages/client/src/components/vectors/ArrowLeft.tsx
@@ -1,12 +1,11 @@
+import * as React from 'react';
 import { SVGProps } from 'react';
-
 const SvgArrowLeft = (props: SVGProps<SVGSVGElement>) => (
-  <svg width="current" height="current" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" {...props}>
     <path
-      d="M20 11v2H8l5.5 5.5-1.42 1.42L4.16 12l7.92-7.92L13.5 5.5 8 11h12Z"
+      d="M19 10.5h-8.793l2.439-2.439a1.5 1.5 0 0 0-2.121-2.122L6.939 9.525a3.505 3.505 0 0 0 0 4.95l3.586 3.586a1.5 1.5 0 0 0 2.121-2.122L10.207 13.5H19a1.5 1.5 0 0 0 0-3Z"
       fill="currentColor"
     />
   </svg>
 );
-
 export default SvgArrowLeft;
