--- packages/client/src/components/vectors/Bookmark.tsx
@@ -1,15 +1,0 @@
-import { SVGProps } from 'react';
-
-const SvgBookmark = (props: SVGProps<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="current"
-    height="current"
-    viewBox="0 0 24 24"
-    {...props}
-  >
-    <path d="M17 3H7a2 2 0 0 0-2 2v16l7-3 7 3V5a2 2 0 0 0-2-2Z" fill="currentColor" />
-  </svg>
-);
-
-export default SvgBookmark;
