+++ packages/client/src/components/vectors/Comment.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import { SVGProps } from 'react';
+const SvgComment = (props: SVGProps<SVGSVGElement>) => (
+  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" {...props}>
+    <path
+      d="M20 0H4a4 4 0 0 0-4 4v12a4 4 0 0 0 4 4h2.9l4.451 3.763a1 1 0 0 0 1.292 0L17.1 20H20a4 4 0 0 0 4-4V4a4 4 0 0 0-4-4ZM7 5h5a1 1 0 0 1 0 2H7a1 1 0 0 1 0-2Zm10 10H7a1 1 0 0 1 0-2h10a1 1 0 0 1 0 2Zm0-4H7a1 1 0 0 1 0-2h10a1 1 0 0 1 0 2Z"
+      fill="currentColor"
+    />
+  </svg>
+);
+export default SvgComment;
