--- packages/client/src/components/vectors/README.md
@@ -1,1 +1,0 @@
-# This is auto generated vector image component by @svgr/cli
