+++ packages/client/src/components/vectors/index.ts
@@ -1,8 +1,8 @@
 export { default as ArrowLeft } from './ArrowLeft';
+export { default as Comment } from './Comment';
+export { default as Create } from './Create';
 export { default as GithubIcon } from './GithubIcon';
-export { default as Bookmark } from './Bookmark';
-export { default as PlusCircle } from './PlusCircle';
-export { default as Search } from './Search';
-export { default as Setting } from './Setting';
 export { default as Home } from './Home';
+export { default as Search } from './Search';
+export { default as Settings } from './Settings';
 export { default as User } from './User';
