--- packages/client/src/pages/Bookmarks/Bookmarks.tsx
@@ -1,7 +1,0 @@
-import TabLayout from '~/components/layouts/TabLayout';
-
-const Bookmarks = () => {
-  return <TabLayout>Bookmarks</TabLayout>;
-};
-
-export default Bookmarks;
