+++ packages/client/src/pages/Home/Home.tsx
@@ -1,55 +1,7 @@
-// hooks
-import { useQueryClient } from '@tanstack/react-query';
-import { useCreateRoom, useGetRooms } from '~/hooks/queries/room';
-
-// components
-import styled from '@emotion/styled';
-import { Button } from '~/components/common';
 import TabLayout from '~/components/layouts/TabLayout';
-import RoomList from './RoomList';
-import AsyncBoundary from '~/components/base/AsyncBoundary';
-import ErrorFallback from '~/components/base/ErrorFallback';
-import { MESSAGE } from '~/constants';
 
 const Home = () => {
-  const queryClient = useQueryClient();
-  const { mutate: createRoom } = useCreateRoom({
-    onSuccess: () => {
-      queryClient.refetchQueries(useGetRooms.getKey());
-    },
-  });
-
-  const handleCreateRoom = () => {
-    createRoom({
-      name: `Room-${crypto.randomUUID()}`,
-    });
-  };
-
-  return (
-    <TabLayout>
-      <Container>
-        <Button shadow onClick={handleCreateRoom}>
-          Create Room
-        </Button>
-        <Spacer />
-        <AsyncBoundary
-          rejectedFallback={
-            <ErrorFallback message={MESSAGE.ERROR.LOAD_DATA} queryKey={useGetRooms.getKey()} />
-          }
-        >
-          <RoomList />
-        </AsyncBoundary>
-      </Container>
-    </TabLayout>
-  );
+  return <TabLayout>Home</TabLayout>;
 };
-
-const Container = styled.div`
-  padding: 16px;
-`;
-
-const Spacer = styled.div`
-  margin: 1rem;
-`;
 
 export default Home;
